Mms: Match count per thread to message per thread
The default SMS messages per thread is 10000, but the max
message counter is set to 5000. This causes the wrong input
number to be shown in the selection dialog, as it's not possible
to go above 5000 and reach the default which is 10000.

Change-Id: I8af30677db2c5c32030ead99f3aa41c5b5469336

(cherry picked from commit 36f220d3ecfbb3a52202e0023c25e8c7adb3b11b)